check pool
